refactored DiskWorker according to the command line design pattern
