+++ jdm.properties
@@ -1,5 +1,5 @@
 #jDiskMark Properties File
-#Fri Oct 23 01:07:19 EDT 2020
+#Wed Dec 02 16:06:11 EST 2020
 numOfFiles=50
 blockSizeKb=128
 autoReset=true
